Adding GameObject and GameObjects classes.
First one represents all kind of inanimate objects like villages, mines, etc. They can be destructible or indestructible, may be marked as a target for the enemy AI. It is implemented in the way that allow to replace easily intact objects with a damaged ones.

Second one is simply an aggregator of GameObject instances.