Pathfinding algorithm added.
It uses BFS to find the shortest path between two points. It works quite well, but has some gotchas: 1) the target tile can not be marked as "blocks" - finder needs to "step on" the final tile to return the path  2) it does not need to step on the first tile, ergo the first time may be blocking  3) but still, the first tile is part of the returned path.  4) to rerun algorithm, the cleanup method must be called.