Player-controlled characters are selectable. Only one player character can be selected at once.
Selected player characters have a separate sprites that are drawn over the basic sprites.

This merge commit adds Beings class that serves as an aggregator of Being instances (both friends and foes), and SpriteTracker class, that is a facade that is going to track and draw all selected entities. Selected entity may be a selected player, targeted enemy, terrain sprite that is part of the path selected by pathfinding algorithm... Currently it works for beings only.