Merging player-movement branch. It allows moving th selectede player from starting to target position (indicated by mouse coursor). It also fixes issue about pathfinding inverting y axis.
To move the player character, user needs to select the player by clicking on the icon (it toggles T/F self.selected for all player controlled characters), move the mouse cursor to the desired position (game instantly draws the path between selected player and cursor) then click on the empty tile (position of player character and its sprite is updated).

Selected player and path between its icon and cursor position are shown as yellow-coloured sprites. Actually the sprite colour is not changed. Instead, SpriteTracker tracks what entities should be marked as active (ie selected, are part of the path, etc.) then draws the new, yellow sprites at the top of the original sprites.