Merging bashing-enemies into development.
This merge adds combat mechanics for player-controlled units.
Beings have attack functionality now. Every attack may consist of various attack effects, and every attack effect has its own pattern and range.
Entities have new type of sprite: "targeted_sprite" that is rendered when player is in the target mode, and the entity is in the range of the attack.
There are also two new game states: "move" when moving mouse over the map will show the path from selected player to cursor, and "target" when all targetable tiles are highlighted and pressing mouse button with perform the attack.

It also fixes #8, and kinda #7 - player actions are modal, but currently it can be toggled by "Enter" key only.

# Conflicts:
#	game/being.py
#	game/beings.py
#	game/game.py