Added new attack type: player can target tile nearby, but only the side tiles of targeted tile will be attacked.
To show how that kind of attacks work, now game in the TARGET mode shows continuously possible targets (in yellow), and when player will mouse over one of these tiles, the area of effect is shown (in red).

Also, current player is now "active" and not selected, because "selected" word is used for targeting.